refactor(react): extract view types to shared viewTypes.ts
Move SortField and SortOrder out of SortControl.tsx into a new
client-react/src/types/viewTypes.ts, and consolidate ViewMode which
was defined locally in three separate files (AppShell, ListViewHeader,
ProjectWorkspaceView). All consumers updated to import from the shared
location.
