+++ client-react/src/types/viewTypes.ts
@@ -1,0 +1,3 @@
+export type SortField = "order" | "createdAt" | "dueDate" | "priority" | "title";
+export type SortOrder = "asc" | "desc";
+export type ViewMode = "list" | "board";
